+++ src/pages/mission/droner/MissionReport.tsx
@@ -88,7 +88,10 @@ function MissionReport() {
         };
         tableList.push(table);
       }
-      const mapPage = res.condition.slice(4 * (current - 1), 4 * current);
+      const mapPage = res.condition.slice(
+        rowCard * (current - 1),
+        rowCard * current
+      );
       setDataInpro(tableList);
       setDataCondition(mapPage);
       setDataMission(res);
@@ -137,7 +140,7 @@ function MissionReport() {
     setCurrentTable(page);
   };
   const onChangePageCard = (page: number) => {
-    setNum(1);
+    setNum(rowCard * (page - 1) + 1);
     setCurrent(page);
   };
 
@@ -150,34 +153,6 @@ function MissionReport() {
       setEndDate(e);
     }
     setCurrent(1);
-  };
-
-  const isNumber = (n: any) => {
-    return !isNaN(parseFloat(n)) && isFinite(n);
-  };
-  const sorter = (a: any, b: any) => {
-    if (a === null) {
-      return 1;
-    }
-    if (b === null) {
-      return -1;
-    }
-    if (isNumber(a) && isNumber(b)) {
-      if (parseInt(a, 10) === parseInt(b, 10)) {
-        return 0;
-      }
-      return parseInt(a, 10) > parseInt(b, 10) ? 1 : -1;
-    }
-    if (isNumber(a)) {
-      return -1;
-    }
-    if (isNumber(b)) {
-      return 1;
-    }
-    if (a === b) {
-      return 0;
-    }
-    return a > b ? 1 : -1;
   };
 
   const mapColor: any = {
@@ -197,7 +172,6 @@ function MissionReport() {
       title: "วันที่อัพเดต",
       dataIndex: "updatedAt",
       key: "updatedAt",
-      sorter: (a: any, b: any) => sorter(a.updatedAt, b.updatedAt),
       render: (value: any, row: any, index: number) => {
         return {
           children: <span>{DateTimeUtil.formatDateTime(row.updateAt)}</span>,
@@ -236,7 +210,6 @@ function MissionReport() {
       title: "จำนวนไร่สะสม",
       dataIndex: "allraiAmount",
       key: "allraiAmount",
-      sorter: (a: any, b: any) => sorter(a.allraiAmount, b.allraiAmount),
       render: (value: any, row: any, index: number) => {
         const calRai = () => {
           let cal = (
@@ -275,7 +248,6 @@ function MissionReport() {
               <span
                 style={{
                   color: color.Success,
-                  fontWeight: "700",
                 }}
               >
                 {row.redeemNo}